requirements.txt
nltk
google-cloud-aiplatform==1.37.0
vertexai==0.0.1
spacy==2.3.5
pyresparser
flask-restful
gunicorn==19.7.1
tomli==2.0.1
pluggy==1.2.0
iniconfig==2.0.0 
pytest==7.4.3
coverage==7.2.6
pytest-cov==2.12.1